Move blk util libraries to sddf/util and remove redefinition of memcpy and memset
