+++ include/sddf/util/bitarray.h
